requirements.txt
# Core Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Web Framework
Django==5.0.1
django-redis==5.4.0
django-influxdb==1.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-django==4.7.0

# API Integration
websockets==11.0.3
aiohttp==3.8.5
python-decouple==3.8

# Development Tools
black==23.7.0
flake8==6.1.0
mypy==1.5.1

# Utils
python-dateutil==2.8.2
pytz==2023.3
